+++ R/tm_g_swimlane.R
@@ -309,14 +309,9 @@ srv_g_swimlane <- function(input, output, session, datasets, dataname,
     anno_txt_var <- input$anno_txt_var
 
     # If reference lines are requested
-    vref_line <- if (!is.null(input$vref_line) && input$vref_line != "") {
-      vref_line <- unlist(strsplit(input$vref_line, ","))
-      vref_line <- as.numeric(vref_line)
-      validate(need(all(!is.na(vref_line)), "Not all values entered for reference line(s) were numeric"))
-      vref_line
-    } else {
-      NULL
-    }
+    vref_line <- as_numeric_from_comma_sep_str(input$vref_line)
+    validate(need(all(!is.na(vref_line)),
+      "Please enter a comma separated set of numeric values for the reference line(s)"))
 
     # validate input values
     if (dataname == "ADSL") {